feat: update job processor

+++ server/src/modules/jobs/crons_actions.ts
@@ -1,9 +1,19 @@
 import cronParser from "cron-parser";
+import { IJobsCron } from "shared/models/job.model";
 
 import logger from "@/common/logger";
 
 import { getDbCollection } from "../../common/utils/mongodbUtils";
-import { createJob, findJob, findJobs, updateJob } from "../actions/job.actions";
+import config from "../../config";
+import {
+  createJobCron,
+  createJobCronTask,
+  createJobSimple,
+  findJobCron,
+  findJobs,
+  updateJob,
+  updateJobCron,
+} from "../actions/job.actions";
 import { CRONS } from "./jobs";
 import { addJob } from "./jobs_actions";
 
@@ -15,52 +25,56 @@ function parseCronString(cronString: string, options: { currentDate: string } | 
 }
 
 export async function cronsInit() {
+  if (config.env === "preview") {
+    return;
+  }
+
   logger.info(`Crons - initialise crons in DB`);
 
-  const crons = Object.values(CRONS);
   let schedulerRequired = false;
 
   await getDbCollection("jobs").deleteMany({
-    name: { $nin: crons.map((c) => c.name) },
+    name: { $nin: CRONS.map((c) => c.name) },
     type: "cron",
   });
   await getDbCollection("jobs").deleteMany({
-    name: { $nin: crons.map((c) => c.name) },
-    status: "pending",
+    name: { $nin: CRONS.map((c) => c.name) },
+    status: { $in: ["pending", "will_start"] },
     type: "cron_task",
   });
 
-  for (const cron of crons) {
-    const cronJob = await findJob({
+  for (const cron of CRONS) {
+    const cronJob = await findJobCron({
       name: cron.name,
-      type: "cron",
     });
 
     if (!cronJob) {
-      await createJob({
+      await createJobCron({
         name: cron.name,
-        type: "cron",
         cron_string: cron.cron_string,
         scheduled_for: new Date(),
       });
       schedulerRequired = true;
-    } else if (cronJob.cron_string !== cron.cron_string) {
-      await updateJob(cronJob._id, {
-        cron_string: cron.cron_string,
+    } else if (cronJob.type === "cron" && cronJob.cron_string !== cron.cron_string) {
+      await updateJobCron(cronJob._id, cron.cron_string);
+      await getDbCollection("jobs").deleteMany({
+        name: cronJob.name,
+        status: { $in: ["pending", "will_start"] },
+        type: "cron_task",
       });
       schedulerRequired = true;
     }
   }
 
   if (schedulerRequired) {
-    await addJob({ name: "crons:scheduler", queued: true });
+    await addJob({ name: "crons:scheduler", queued: true, payload: {} });
   }
 }
 
 export async function cronsScheduler(): Promise<void> {
   logger.info(`Crons - Check and run crons`);
 
-  const crons = await findJobs(
+  const crons = await findJobs<IJobsCron>(
     {
       type: "cron",
       scheduled_for: { $lte: new Date() },
@@ -72,8 +86,7 @@ export async function cronsScheduler(): Promise<void> {
     const next = parseCronString(cron.cron_string ?? "", {
       currentDate: cron.scheduled_for,
     }).next();
-    await createJob({
-      type: "cron_task",
+    await createJobCronTask({
       name: cron.name,
       scheduled_for: next.toDate(),
     });
@@ -82,19 +95,15 @@ export async function cronsScheduler(): Promise<void> {
       scheduled_for: next.toDate(),
     });
   }
-  const cron = await findJob(
-    {
-      type: "cron",
-    },
-    { sort: { scheduled_for: 1 } }
-  );
+  const cron = await findJobCron({}, { sort: { scheduled_for: 1 } });
 
   if (!cron) return;
 
   cron.scheduled_for.setSeconds(cron.scheduled_for.getSeconds() + 1); // add DELTA of 1 sec
-  await createJob({
-    type: "simple",
+  await createJobSimple({
     name: "crons:scheduler",
     scheduled_for: cron.scheduled_for,
+    sync: false,
+    payload: {},
   });
 }
